Implemented basic player movement
Created PlayerMovement.cs Implemented a way to move in the 4 cardinal directions. Also implemented a very primitive dash function (the player just moves much faster)

Edited PlayerController.cs to have a getter and setter for speed.

Not tested because there is no player yet, but should work with minimal changes.